requirements.txt
phidata
python-dotenv
yfinance
packaging 
duckduckgo-search
fastapi 
uvicorn 
groq
openai
youtube_transcript_api
requests 
beautifulsoup4 
langchain 
chromadb 
tiktoken
langchain-community
sentence-transformers
langchain-openai
ollama
langchain-ollama
sqlalchemy
psycopg[binary]
pgvector
pypdf